Revert: Fix panic when trying to format max u16 as signed hex

+++ src/ast.rs
@@ -101,9 +101,10 @@ impl Immediate {
 struct Signed(u16);
 // Format an i16 as a sign-aware hex string
 // https://stackoverflow.com/a/44712309
+// TODO: fix panic when formatting max u16 as signed int
 impl fmt::LowerHex for Signed {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let val = self.0 as i32;
+        let val = self.0 as i16;
         let p = if f.alternate() { "0x" } else { "" };
         let x = format!("{:x}", val.abs());
         f.pad_integral(val >= 0, p, &x)
